refactor(plugin-sdk): remove final test-only facades
