+++ config/knip.config.ts
@@ -683,7 +683,6 @@ const config = {
       "browser-control-auth.ts!",
       "browser-config.ts!",
       "browser-doctor.ts!",
-      "browser-host-inspection.ts!",
       "browser-maintenance.ts!",
       "browser-profiles.ts!",
       // Built by tsdown as the native messaging executable; Chrome launches it by path.
