--- src/plugin-sdk/google-model-id.ts
@@ -1,7 +1,0 @@
-/**
- * Public SDK subpath for normalizing Google and Antigravity preview model ids.
- */
-export {
-  normalizeAntigravityPreviewModelId as normalizeAntigravityModelId,
-  normalizeGooglePreviewModelId as normalizeGoogleModelId,
-} from "./provider-model-shared.js";
